Add link to client journey list, format date, etc. for Jane and client tables
