+++ react-app/package.json
@@ -23,6 +23,7 @@
     "firebase": "^11.3.1",
     "html-to-image": "^1.11.13",
     "lucide-react": "^0.503.0",
+    "luxon": "^3.6.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.5.0",
@@ -37,6 +38,7 @@
   "devDependencies": {
     "@eslint/js": "^9.19.0",
     "@types/downloadjs": "^1.4.6",
+    "@types/luxon": "^3.6.2",
     "@types/node": "^22.15.0",
     "@types/react": "^18.2.55",
     "@types/react-dom": "^18.2.19",
